+++ pallets/sidechain/src/lib.rs
@@ -6,12 +6,28 @@ pub mod mock;
 mod tests;
 
 pub use pallet::*;
+use sidechain_domain::McBlockHash;
+use sp_core::{Decode, Encode};
+use sp_inherents::IsFatalError;
+
+#[derive(Encode, sp_runtime::RuntimeDebug, PartialEq, Eq)]
+#[cfg_attr(feature = "std", derive(Decode))]
+pub enum InherentError {
+	DoesNotMatchInherentData { expected: McBlockHash, actual: McBlockHash },
+	ChangeNotHandled,
+}
+
+impl IsFatalError for InherentError {
+	fn is_fatal_error(&self) -> bool {
+		true
+	}
+}
 
 #[frame_support::pallet]
 pub mod pallet {
 	use frame_support::pallet_prelude::*;
-	use frame_system::pallet_prelude::BlockNumberFor;
-	use sidechain_domain::{ScEpochNumber, ScSlotNumber};
+	use frame_system::pallet_prelude::*;
+	use sidechain_domain::{McBlockHash, ScEpochNumber, ScSlotNumber};
 	use sp_sidechain::OnNewEpoch;
 
 	#[pallet::pallet]
@@ -39,6 +55,11 @@ pub mod pallet {
 	#[pallet::getter(fn slots_per_epoch)]
 	pub(super) type SlotsPerEpoch<T: Config> =
 		StorageValue<_, sidechain_slots::SlotsPerEpoch, ValueQuery>;
+
+	#[pallet::storage]
+	#[pallet::getter(fn last_mc_hash)]
+	pub(super) type LastMcHash<T: Config> =
+		StorageValue<_, sidechain_domain::McBlockHash, OptionQuery>;
 
 	#[pallet::storage]
 	pub(super) type SidechainParams<T: Config> = StorageValue<_, T::SidechainParams, ValueQuery>;
@@ -69,6 +90,75 @@ pub mod pallet {
 		fn build(&self) {
 			SidechainParams::<T>::put(self.params.clone());
 			SlotsPerEpoch::<T>::put(self.slots_per_epoch);
+		}
+	}
+
+	#[pallet::inherent]
+	impl<T: Config> ProvideInherent for Pallet<T> {
+		type Call = Call<T>;
+		type Error = crate::InherentError;
+		const INHERENT_IDENTIFIER: InherentIdentifier = sidechain_mc_hash::INHERENT_IDENTIFIER;
+
+		/// Responsible for calling `Call:set()` on each block by the block author, if the validator list changed
+		fn create_inherent(data: &InherentData) -> Option<Self::Call> {
+			let mc_hash = Self::get_mc_hash_from_inherent_data(data);
+
+			match LastMcHash::<T>::get() {
+				None => Some(Call::set_last_mc_hash { mc_hash }),
+				Some(last_mc_hash) if last_mc_hash == mc_hash => None,
+				Some(_) => Some(Call::set_last_mc_hash { mc_hash }),
+			}
+		}
+
+		fn check_inherent(call: &Self::Call, data: &InherentData) -> Result<(), Self::Error> {
+			if let Self::Call::set_last_mc_hash { mc_hash: expected } = call {
+				let actual = Self::get_mc_hash_from_inherent_data(data);
+				if *expected != actual {
+					return Err(Self::Error::DoesNotMatchInherentData {
+						expected: expected.clone(),
+						actual,
+					});
+				}
+			}
+
+			Ok(())
+		}
+
+		fn is_inherent(call: &Self::Call) -> bool {
+			matches!(call, Call::set_last_mc_hash { .. })
+		}
+
+		fn is_inherent_required(data: &InherentData) -> Result<Option<Self::Error>, Self::Error> {
+			let mc_hash = Self::get_mc_hash_from_inherent_data(data);
+			match LastMcHash::<T>::get() {
+				Some(last_mc_hash) if last_mc_hash != mc_hash => {
+					Ok(Some(Self::Error::ChangeNotHandled))
+				},
+				_ => Ok(None),
+			}
+		}
+	}
+
+	impl<T: Config> Pallet<T> {
+		fn get_mc_hash_from_inherent_data(data: &InherentData) -> sidechain_domain::McBlockHash {
+			data.get_data::<McBlockHash>(&Self::INHERENT_IDENTIFIER)
+				.expect("❌ MC Block Hash inherent data is invalid")
+				.expect("❌ MC Block Hash inherent data is missing")
+		}
+	}
+
+	#[pallet::call]
+	impl<T: Config> Pallet<T> {
+		#[pallet::call_index(0)]
+		#[pallet::weight((1, DispatchClass::Mandatory))]
+		pub fn set_last_mc_hash(
+			origin: OriginFor<T>,
+			mc_hash: sidechain_domain::McBlockHash,
+		) -> DispatchResult {
+			ensure_none(origin)?;
+			log::info!("#️⃣ New MC block referenced: {mc_hash}");
+			LastMcHash::<T>::put(mc_hash);
+			Ok(())
 		}
 	}